requirements.txt
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.1
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
crashtest==0.3.1
cryptography==38.0.1
distlib==0.3.6
dnspython==2.2.1
dulwich==0.20.45
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.81.0
filelock==3.8.0
greenlet==1.1.3
h11==0.13.0
html5lib==1.1
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
jaraco.classes==3.2.2
Jinja2==3.1.2
jsonschema==4.15.0
keyring==23.9.0
lockfile==0.12.2
Mako==1.2.3
MarkupSafe==2.1.1
more-itertools==8.14.0
msgpack==1.0.4
orjson==3.8.0
packaging==21.3
passlib==1.7.4
pexpect==4.8.0
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.2.0
poetry-core==1.1.0
poetry-plugin-export==1.0.6
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.1
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
tomlkit==0.11.4
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.16.4
watchgod==0.8.2
webencodings==0.5.1
websockets==10.3
xattr==0.9.9
